refactor: extract popularity types
