+++ src/components/modals/ClassInfo/ClassInfoModal.tsx
@@ -3,7 +3,8 @@ import React, { Dispatch, SetStateAction } from "react";
 
 import ClassInfo from "@/components/modals/ClassInfo/ClassInfo";
 import { classRecurrentId } from "@/lib/helpers/recurrentId";
-import { AllConfigsIndex, ClassPopularity, ClassPopularityIndex, RezervoClass } from "@/types/rezervo";
+import { ClassPopularity, ClassPopularityIndex } from "@/types/popularity";
+import { AllConfigsIndex, RezervoClass } from "@/types/rezervo";
 
 const ClassInfoModal = ({
     classInfoClass,
