+++ src/components/schedule/DaySchedule.tsx
@@ -5,13 +5,8 @@ import React from "react";
 import ClassCard from "@/components/schedule/class/ClassCard";
 import { getCapitalizedWeekday } from "@/lib/helpers/date";
 import { classRecurrentId } from "@/lib/helpers/recurrentId";
-import {
-    AllConfigsIndex,
-    ClassPopularity,
-    ClassPopularityIndex,
-    RezervoClass,
-    RezervoDaySchedule,
-} from "@/types/rezervo";
+import { ClassPopularity, ClassPopularityIndex } from "@/types/popularity";
+import { AllConfigsIndex, RezervoClass, RezervoDaySchedule } from "@/types/rezervo";
 
 function DaySchedule({
     daySchedule,
