+++ src/components/schedule/WeekSchedule.tsx
@@ -3,7 +3,8 @@ import { useRouter } from "next/router";
 import React, { useEffect } from "react";
 
 import DaySchedule from "@/components/schedule/DaySchedule";
-import { AllConfigsIndex, ClassPopularityIndex, RezervoClass, RezervoWeekSchedule } from "@/types/rezervo";
+import { ClassPopularityIndex } from "@/types/popularity";
+import { AllConfigsIndex, RezervoClass, RezervoWeekSchedule } from "@/types/rezervo";
 
 function WeekSchedule({
     weekSchedule,
