+++ src/types/rezervo.ts
@@ -71,17 +71,6 @@ export type UserSessionsIndex = {
     [classId: string]: UserNameSessionStatus[];
 };
 
-export enum ClassPopularity {
-    Unknown = "Denne timen gikk ikke forrige uke, og har derfor ukjent popularitet.",
-    Low = "Denne timen har vanligvis mange ledige plasser.",
-    Medium = "Denne timen har vanligvis noen ledige plasser.",
-    High = "Denne timen er vanligvis full.",
-}
-
-export type ClassPopularityIndex = {
-    [recurrentId: string]: ClassPopularity;
-};
-
 export type IntegrationProfile = {
     acronym: IntegrationIdentifier;
     name: string;
